scripts/requirements-query.txt
# Requirements for Supabase SQL Query Runner
psycopg2-binary>=2.9.0
supabase>=2.0.0
requests>=2.31.0

